feat: support private registry host (#28)
This pull request adds support for non-default docker hosts (such as self-hosted registries or services like Amazon ECR). It introduces an optional environment variable DOCKER_REGISTRY to denote the URL of the registry. If this variable is left blank, the default registry of the docker client (probably Docker Hub) will be used.