feat(commitmatic): adds new ai powered commit utility & cli
* wip

* feat: add support for custom commit message

Add support for generating custom commit messages using the commitmatic tool. This allows users to define their own commit message convention and generate clear and compliant commit messages based on the information they provide. The commitmatic tool provides a helpful assistant for crafting commit messages.

* chore(dev-deps): update husky dependency

- Update husky dependency from "7.0.4" to "^7.0.4"

* feat(deps): adds dedent dependency

Add the `dedent` dependency to the `package.json` files in the root directory and the `packages/commitmatic` directory.

* feat(commitmatic): adds shebang to index.ts file

Add a shebang line to the `index.ts` file in the `commitmatic` package. This will ensure that the file can be executed as a standalone script.

* feat(commitmatic): add initial spec for commitmatic

This commit adds the initial spec for the commitmatic library, which is responsible for generating conventional commit messages based on the user's input. This spec defines the expected behavior of the commitmatic function.
test: remove invalid test

---------

Co-authored-by: JoA-MoS <JoA-MoS@users.noreply.github.com>